Move all legacy steps to shared file
With all our clean up of the legacy features the duplication across them has become more acute. There is almost no difference between them.

We've had to make some changes to support making _all_ steps regime independent, most notably adding a table for retrospective transactions and updating the column information.

But with these changes all 3 legacy features now use the same steps. There is still a lot of duplication between them and existing steps. Plus they legacy steps should be updated to use our page objects. But for now this gives us a base to build on.
